add streaming status to device
