+++ ui/test/ex_nvr_web/controllers/api/system_status_controller_test.exs
@@ -1,4 +1,4 @@
-defmodule ExNVRWeb.API.DeviceControllerTest do
+defmodule ExNVRWeb.API.SystemStatusControllerTest do
   @moduledoc false
 
   use ExNVRWeb.ConnCase
